Revert "delta cascade"
This reverts commit b839b3dbd3e11a8ea743aae0ded641313dc12c61.

+++ src/highlevel/CascadedMetadata.h
@@ -97,7 +97,7 @@ public:
 
   struct Header
   {
-    uint32_t length; // !!! change #define OPTIMIZATION_TASK
+    uint64_t length;
     MinValue minValue;
     uint8_t numBits;
   };
